separate ReservationTab from MainActivity
